updated task3.py with csv creation code + organization
