+++ task3/old/lmqltest.py
