+++ task3/task3_run.sh
@@ -1,0 +1,4 @@
+for filename in *.json; do
+    python3 task3.py "$filename"
+
+done
